FIX : scss 구조 수정
