+++ itzi/src/assets/scss/section/_footer.scss
@@ -1,0 +1,27 @@
+.footer{
+    .container{
+        width:1440px;
+        height:118px;
+        background-color:#6E6F88;
+        display:flex;
+        flex-direction: column;
+        justify-content: center;
+    }
+
+    .top{
+        display:flex;
+        gap:18.95px;
+        color:#D9DAEC;
+        font-size:12px;
+        font-weight:500;
+        justify-content: flex-end;
+        padding-right:74.27px;
+    }
+
+    .bottom{
+        display:flex;
+        gap:14.53px;
+        justify-content: flex-end;
+        padding-right:74.27px;
+    }
+}
