+++ itzi/src/assets/scss/section/_main.scss
@@ -8,5 +8,5 @@
 .body{
     width:1440px;
     margin:0 auto;
-    border:1px solid red;
+    border:1px solid red; // 영역 구분하기 위해서 설정해놓음
 }
