--- itzi/src/assets/scss/section/nav.module.scss
@@ -1,44 +1,0 @@
-.container{
-    height:51px;
-    display:flex;
-    align-items: center;
-    padding:0 83.1px;
-    justify-content: space-between;
-}
-.nav_left{
-    display:flex;
-    gap:38.4px;
-}
-.nav_right{
-    display:flex;
-    align-items: center;
-}
-.banner{
-    display:flex;
-    align-items: center;
-}
-.button_container{
-    display:flex;
-    gap:22px;
-}
-.button_container>button{
-    background: none;
-    border: none;
-    -webkit-appearance: none;
-    appearance: none;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 600;
-    color:#3A3B4D;
-}
-.button_container>p{
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 600;
-    color:#3A3B4D;
-}
-.icon_container{
-    gap:19.23px;
-    display:flex;
-    align-items: center;
-}
